moved index folder around
