+++ .husky/pre-commit
@@ -1,5 +1,5 @@
 #!/usr/bin/env sh
-. "$(dirname -- "$0")/_/husky.sh"
+."$(dirname -- "$0")/_/husky.sh"
 
 yarn lint-staged
 yarn test
